+++ src/capacitor-open.ts
@@ -14,6 +14,7 @@ export function capacitorOpen(project: Project, platform: CapacitorPlatform): st
 	const ionicCLI = exists('@ionic/cli');
 	switch (project.repoType) {
 		case MonoRepoType.none: return ionicCLI ? capCLIOpen(platform) : ionicCLIOpen(platform);
+		case MonoRepoType.folder:
 		case MonoRepoType.npm: return InternalCommand.cwd + (ionicCLI ? capCLIOpen(platform) : ionicCLIOpen(platform));
 		case MonoRepoType.nx: return nxOpen(project, platform);
 		default: throw new Error('Unsupported Monorepo type');